+++ CHANGELOG.md
@@ -1,1 +1,6 @@
-- Added more events
+- Added more events.
+- Allow binding keybind config to panel visibility.
+- Added `clearpickups` command (removes all pickups in a radius).
+- Fixed some inputs not working sometimes.
+- Various bug fixes.
+- SUI improvements.